Rename interface Test to App and move subprojects to root
