+++ build.gradle
@@ -1,12 +1,10 @@
-version = "0.1.0"
-
 allprojects {
   apply plugin: 'idea'
   apply plugin: 'eclipse'
   apply plugin: 'java'
   apply plugin: 'checkstyle'
 
-  version = "0.1.0"
+  version = "1.0.0"
 
   sourceCompatibility = 1.7
 
@@ -14,22 +12,6 @@ allprojects {
     mavenCentral()
   }
 
-  checkstyle {
-    configFile = new File(rootDir, "checkstyle/checkstyle.xml")
-  }
-  test.dependsOn('checkstyleMain', 'checkstyleTest')
-
-  test {
-    useTestNG()
-
-    testLogging {
-      events "passed", "skipped", "failed"
-      exceptionFormat = 'full'
-    }
-  }
-}
-
-project(':services') {
   dependencies {
     compile 'net.sourceforge.argparse4j:argparse4j:0.5.0'
     compile 'org.slf4j:slf4j-log4j12:1.7.6'
@@ -68,13 +50,11 @@ project(':services') {
     archives testJar
   }
 
-}
-
-project(':tests') {
-
-  dependencies {
-    compile project(':services')
+  checkstyle {
+    configFile = new File(rootDir, "checkstyle/checkstyle.xml")
   }
+  
+  test.dependsOn('checkstyleMain', 'checkstyleTest')
 
   test {
     useTestNG()
@@ -85,13 +65,6 @@ project(':tests') {
     }
   }
 }
-
-project(':core') {
-  dependencies {
-    compile project(':tests')
-  }
-}
-
 
 task wrapper(type: Wrapper) {
    gradleVersion = '2.11'
