+++ config/kafka-monitor.properties
@@ -34,7 +34,7 @@
 
 {
   "end-to-end": {
-    "class.name": "com.linkedin.kmf.tests.BasicEndToEndTest",
+    "class.name": "com.linkedin.kmf.apps.SingleClusterMonitor",
     "topic": "kafka-monitor-topic",
     "zookeeper.connect": "localhost:2181",
     "bootstrap.servers": "localhost:9092",
