--- settings.gradle
@@ -1,3 +1,0 @@
-include 'core'
-include 'services'
-include 'tests'
