+++ src/main/java/com/linkedin/kmf/consumer/BaseConsumerRecord.java
