+++ src/main/java/com/linkedin/kmf/consumer/KMBaseConsumer.java
