+++ src/main/java/com/linkedin/kmf/consumer/NewConsumer.java
